pyproject.toml
[project]
name = "pynpxpipe"
version = "0.1.0"
description = "Neural electrophysiology preprocessing pipeline for SpikeGLX multi-probe recordings"
readme = "README.md"
authors = [
    { name = "GongZhengxin", email = "1045418215@qq.com" }
]
requires-python = ">=3.11,<3.13"
dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    # Electrophysiology pipeline
    "spikeinterface[full]>=0.104.0",
    "probeinterface>=0.2.0",
    # NWB output
    "pynwb>=2.8.0",
    "neo>=0.13.0",
    # Behavioral data parsing
    "h5py>=3.8.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",    # required by pandas for .parquet read/write
    # Configuration
    "pyyaml>=6.0.0",
    # Structured logging
    "structlog>=23.0.0",
    # CLI framework (only used in cli/ layer)
    "click>=8.1.0",
    # Resource detection
    "psutil>=5.9.0",            # CPU / RAM / Disk detection (required)
]

[project.optional-dependencies]
gpu = [
    "nvidia-ml-py>=12.535.0",  # pynvml: GPU VRAM detection via NVML (optional, Windows/Linux)
]
matlab = [
    "matlabengine>=23.2",      # MATLAB Engine for Python (BHV2 parsing, M1 only)
]
plots = [
    "matplotlib>=3.7",         # Sync diagnostic plots (optional)
]
ui = [
    "panel>=1.0",              # Web UI (optional)
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    # Tutorial notebooks
    "jupyterlab>=4.0.0",
    "nbmake>=1.4.0",
]

[project.scripts]
pynpxpipe = "pynpxpipe.cli.main:cli"
pynpxpipe-ui = "pynpxpipe.ui.app:main"

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["pynpxpipe"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
